Dynamic Subnav List Loading for Marketing Solutions Main Page, finished scrollhandler
The Subnav List will load in dynamically (including the height for each) depending on if it's a Marketing Solution or Our Clients Page.
This is onlye setup for the main page of Marketing Solutions since the component is only on that page. The OnScrollListener that is
used to stick the subnav to the bottom of the navigation is also removed when the component is unmounted so that it doesn't cause errors
on subsequent pages.